+++ midolman/src/test/scala/org/midonet/midolman/DeduplicationActorTestCase.scala
@@ -86,8 +86,8 @@ class DeduplicationActorTestCase extends MidolmanSpec {
             override def host: Host = new Host(UUID.randomUUID(), true, 0, "",
                                                Map(), Map())
             override def peerTunnelInfo(peer: UUID): Option[Route] = ???
-            override def isVtepTunnellingPort(portNumber: Short): Boolean = ???
-            override def isOverlayTunnellingPort(portNumber: Short): Boolean = ???
+            override def isVtepTunnellingPort(portNumber: Integer): Boolean = ???
+            override def isOverlayTunnellingPort(portNumber: Integer): Boolean = ???
             override def vtepTunnellingOutputAction: FlowActionOutput = ???
         })
         dda.hookPacketHandler()
